refactor: move MultiHeadAttention4D to timeseries_models and remove old implementation
