+++ hephaestus/timeseries_models/model_data_classes.py
